Turn debug mode off by default
This will load the version of the module that was compiled with
release optimizations, rather than the one with debug symbols.

This is useful as I will be starting soon with benchmarking, where the
overhead from the debug symbols and lack of optimizations may or may
not have a noticeable performance impact.